legion/requirements/base.txt
dill
python-interface
pandas
traitlets
docker
python-consul
flask
numpy
Pillow
Jinja2>=2.9
statsd
kubernetes
jsonschema
PyYAML